+++ source/extensions/isaacsim.replicator.behavior/isaacsim/replicator/behavior/behaviors/example_behavior.py
@@ -37,8 +37,8 @@ class ExampleBehavior(BehaviorScript):
         {
             "attr_name": "includeChildren",
             "attr_type": Sdf.ValueTypeNames.Bool,
-            "default_value": False,
-            "doc": "Include prim children in the behavior.",
+            "default_value": True,
+            "doc": "Include valid prim children to the behavior.",
         },
     ]